Finished importing all the data from the files.
