+++ src/example-np/test_DATA.dat
@@ -1,0 +1,35 @@
+     NODE COORDINATES
+----------------------------------------
+NODE       X(M)        Y(M)
+1          0.000       0.000
+2          0.000       2.000
+3          0.000       4.000
+4          4.000       4.000
+5          4.000       2.000
+
+ ELEMENTS
+----------------------------------------
+EL.  NODE1  NODE2      A(M2)     E(PA)
+1   4      3                1      2E+11
+2   3      2                1      2E+11
+3   2      4                1      2E+11
+4   4      5                1      2E+11
+5   5      2                1      2E+11
+6   2      1                1      2E+11
+7   1      5                1      2E+11
+
+  PIN SUPPORTS
+----------------------------------------
+NODE    DX'(M)    DY'(M)   ANGLE(DEG)
+4          0.000       0.000        0.00
+
+ROLLER SUPPORTS
+----------------------------------------
+NODE   DIRECTION   DN(M)   ANGLE(DEG)
+1    2.0               0.000       26.57
+
+ FORCES
+----------------------------------------
+NODE     FX'(N)     FY'(N)   ANGLE(DEG)
+5          40000           0     -180.00
+3          20000           0      200.00
